refactor: app 폴더 내 hooks를 component 폴더로 이동
