+++ src/_components/Home/_utils/validateEmail.ts
